Finished moving Section 6 into it's own file; Moving Section 5 keys into separate file (#24907)
