+++ modules/pensions/lib/pensions/pdf_fill/sections/section_06.rb
@@ -176,7 +176,61 @@ module Pensions
       # Note: This method modifies `form_data`
       #
       def expand(form_data)
-        # Add expansion logic here
+        form_data['maritalStatus'] = marital_status_to_radio(form_data['maritalStatus'])
+        form_data['currentMarriage'] = get_current_marriage(form_data['marriages'])
+        form_data['spouseDateOfBirth'] = split_date(form_data['spouseDateOfBirth'])
+        form_data['spouseSocialSecurityNumber'] = split_ssn(form_data['spouseSocialSecurityNumber'])
+        form_data['spouseIsVeteran'] = to_radio_yes_no(form_data['spouseIsVeteran']) if form_data['maritalStatus'] != 2
+        form_data['spouseAddress'] ||= {}
+        form_data['spouseAddress']['postalCode'] = split_postal_code(form_data['spouseAddress'])
+        form_data['spouseAddress']['country'] = form_data.dig('spouseAddress', 'country')&.slice(0, 2)
+        form_data['currentSpouseMonthlySupport'] =
+          split_currency_amount(form_data['currentSpouseMonthlySupport'])
+        form_data['reasonForCurrentSeparation'] =
+          reason_for_current_separation_to_radio(form_data['reasonForCurrentSeparation'])
+      end
+
+      # Take a marital status and convert it to a radio selection.
+      def marital_status_to_radio(marital_status)
+        case marital_status
+        when 'MARRIED' then 0
+        when 'SEPARATED' then 1
+        else 2
+        end
+      end
+
+      # Get the current marriage
+      def get_current_marriage(marriages)
+        current_marriage_index = marriages&.index { |marriage| !marriage.key?('dateOfSeparation') }
+
+        if current_marriage_index
+          current_marriage = marriages[current_marriage_index].clone
+          marriages.delete_at(current_marriage_index)
+        else
+          current_marriage = {}
+        end
+
+        return current_marriage if current_marriage.empty?
+
+        middle_initial = current_marriage.dig('spouseFullName', 'middle')&.first
+        current_marriage['spouseFullName']['middle'] = middle_initial
+        marriage_type = current_marriage['marriageType']
+        current_marriage['marriageType'] =
+          marriage_type == 'CEREMONY' ? 0 : 1
+        current_marriage['dateOfMarriage'] =
+          split_date(current_marriage['dateOfMarriage'])
+        current_marriage
+      end
+
+      # Get the current reason of separation to a radio box.
+      def reason_for_current_separation_to_radio(reason_for_separation)
+        case reason_for_separation
+        when 'MEDICAL_CARE' then 0
+        when 'RELATIONSHIP' then 1
+        when 'LOCATION' then 2
+        when 'OTHER' then 3
+        else 'Off'
+        end
       end
     end
   end
