+++ modules/pensions/spec/lib/pensions/pdf_fill/va21p527ez_spec.rb
@@ -77,24 +77,6 @@ describe Pensions::PdfFill::Va21p527ez do
     end
   end
 
-  describe '#marital_status_to_radio' do
-    it 'returns correct radio value for marital status' do
-      expect(described_class.new({}).marital_status_to_radio('MARRIED')).to eq(0)
-      expect(described_class.new({}).marital_status_to_radio('SEPARATED')).to eq(1)
-      expect(described_class.new({}).marital_status_to_radio('SINGLE')).to eq(2)
-    end
-  end
-
-  describe '#reason_for_current_separation_to_radio' do
-    it 'returns correct radio value for current separation reasons' do
-      expect(described_class.new({}).reason_for_current_separation_to_radio('MEDICAL_CARE')).to eq(0)
-      expect(described_class.new({}).reason_for_current_separation_to_radio('RELATIONSHIP')).to eq(1)
-      expect(described_class.new({}).reason_for_current_separation_to_radio('LOCATION')).to eq(2)
-      expect(described_class.new({}).reason_for_current_separation_to_radio('OTHER')).to eq(3)
-      expect(described_class.new({}).reason_for_current_separation_to_radio('')).to eq('Off')
-    end
-  end
-
   describe '#expand_veteran_service_information' do
     it 'puts overflow on line one' do
       long_place_of_separation = 'A very long place name that exceeds thirty-six characters'
